# apps/api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "Intuitive Fusion Procurement Portal API (POC)",
  "version": "0.1.0-draft",
  "description": "REST API for the POC. Auth is mock (HttpOnly session cookie) behind an IdentityProvider interface. All mutations emit audit events. Errors follow RFC 7807 (application/problem+json) with stable `code` and `correlationId`."
 },
 "servers": [
  {
   "url": "http://localhost:4000/api/v1",
   "description": "Local dev"
  }
 ],
...
 "security": [
  {
   "sessionCookie": []
  }
 ],
 "paths": {
...
  "/admin/users": {
   "get": {
    "operationId": "adminListUsers",
    "tags": [
     "Admin"
    ],
    "summary": "List staff users with their roles (read only; suppliers are not listed)",
    "parameters": [],
    "x-roles": [
     "ADMIN"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AdminUser"
         }
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    }
   },
   "post": {
    "operationId": "adminCreateUser",
    "tags": [
     "Admin"
    ],
    "summary": "Create user",
    "parameters": [],
    "x-roles": [
     "ADMIN"
    ],
    "responses": {
     "201": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AdminUser"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AdminUserCreate"
       }
      }
     }
    }
   }
  },
  "/admin/delegations": {
   "get": {
    "operationId": "listDelegations",
    "tags": [
     "Admin"
    ],
    "summary": "Delegations of authority",
    "parameters": [],
    "x-roles": [
     "ADMIN",
     "EXEC"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Delegation"
         }
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    }
   },
   "post": {
    "operationId": "createDelegation",
    "tags": [
     "Admin"
    ],
    "summary": "Grant a limit to a person (or role); audited; applies to the next approval or signature",
    "parameters": [],
    "x-roles": [
     "ADMIN"
    ],
    "responses": {
     "201": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Delegation"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "description": "409 DELEGATION_EXISTS; 403 for oneself",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DelegationCreate"
       }
      }
     }
    }
   }
  },
  "/admin/delegations/{id}": {
   "put": {
    "operationId": "updateDelegation",
    "tags": [
     "Admin"
    ],
    "summary": "Change a threshold or switch it off; audited; effective immediately; the person is notified",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "ADMIN"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Delegation"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DelegationUpdate"
       }
      }
     }
    }
   }
  },
  "/admin/alert-settings": {
   "get": {
    "operationId": "getAlertSettings",
    "tags": [
     "Admin"
    ],
    "summary": "Contract alert lead times in days",
    "parameters": [],
    "x-roles": [
     "ADMIN"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AlertSettings"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    }
   },
   "put": {
    "operationId": "setAlertSettings",
    "tags": [
     "Admin"
    ],
    "summary": "Change the lead times; scheduled alerts of executed contracts move at once; audited",
    "parameters": [],
    "x-roles": [
     "ADMIN"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AlertSettings"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AlertSettings"
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "sessionCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "if_session"
   }
  },
  "schemas": {
   "Problem": {
    "type": "object",
    "properties": {
     "type": {
      "type": "string"
     },
     "title": {
      "type": "string"
     },
     "status": {
      "type": "integer"
     },
     "detail": {
      "type": "string"
     },
     "code": {
      "type": "string"
     },
     "correlationId": {
      "type": "string"
     },
     "errors": {
      "type": "array",
      "items": {
       "type": "object",
       "properties": {
        "field": {
         "type": "string"
        },
        "message": {
         "type": "string"
        }
       },
       "additionalProperties": false,
       "required": [
        "field",
        "message"
       ]
      }
     }
    },
    "additionalProperties": false,
    "required": [
     "type",
     "title",
     "status",
     "code",
     "correlationId"
    ]
   },
...
   "Delegation": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "userId": {
      "type": "string",
      "format": "uuid"
     },
     "userName": {
      "type": "string"
     },
     "role": {
      "type": "string"
     },
     "scope": {
      "type": "string",
      "enum": [
       "SOURCING_APPROVAL",
       "CONTRACT_SIGNING",
       "PUBLISH_PERMISSION"
      ]
     },
     "maxValue": {
      "type": "number"
     },
     "division": {
      "type": "string"
     },
     "active": {
      "type": "boolean"
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "scope",
     "maxValue"
    ]
   },
   "DelegationUpdate": {
    "type": "object",
    "properties": {
     "maxValue": {
      "type": "number",
      "minimum": 0
     },
     "active": {
      "type": "boolean"
     }
    },
    "additionalProperties": false,
    "required": [
     "maxValue"
    ]
   },
   "AdminUser": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "name": {
      "type": "string"
     },
     "email": {
      "type": "string"
     },
     "roles": {
      "type": "array",
      "items": {
       "type": "string",
       "enum": [
        "REQUESTER",
        "PROCUREMENT",
        "DELEGATE",
        "EVALUATOR",
        "CHAIR",
        "LEGAL",
        "CONTRACT_MGR",
        "PROBITY",
        "FINANCE",
        "ADMIN",
        "EXEC",
        "SUPPLIER"
       ]
      }
     },
     "orgUnit": {
      "type": "string"
     },
     "active": {
      "type": "boolean"
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "name",
     "email",
     "roles"
    ]
   },
   "DelegationCreate": {
    "type": "object",
    "properties": {
     "scope": {
      "type": "string",
      "enum": [
       "SOURCING_APPROVAL",
       "CONTRACT_SIGNING",
       "PUBLISH_PERMISSION"
      ]
     },
     "userId": {
      "type": "string",
      "format": "uuid"
     },
     "role": {
      "type": "string",
      "enum": [
       "REQUESTER",
       "PROCUREMENT",
       "DELEGATE",
       "EVALUATOR",
       "CHAIR",
       "LEGAL",
       "CONTRACT_MGR",
       "PROBITY",
       "FINANCE",
       "ADMIN",
       "EXEC",
       "SUPPLIER"
      ]
     },
     "maxValue": {
      "type": "number",
      "minimum": 0
     },
     "division": {
      "type": "string"
     }
    },
    "additionalProperties": false,
    "required": [
     "scope",
     "maxValue"
    ]
   },
   "AlertSettings": {
    "type": "object",
    "properties": {
     "expiry": {
      "type": "integer",
      "minimum": 1,
      "maximum": 365
     },
     "notice": {
      "type": "integer",
      "minimum": 1,
      "maximum": 365
     },
     "extension": {
      "type": "integer",
      "minimum": 1,
      "maximum": 365
     },
     "milestone": {
      "type": "integer",
      "minimum": 1,
      "maximum": 365
     }
    },
    "additionalProperties": false,
    "required": [
     "expiry",
     "notice",
     "extension",
     "milestone"
    ]
   },
...
   "AdminUserCreate": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "email": {
      "type": "string",
      "format": "email"
     },
     "role": {
      "type": "string",
      "enum": [
       "REQUESTER",
       "PROCUREMENT",
       "DELEGATE",
       "EVALUATOR",
       "CHAIR",
       "LEGAL",
       "CONTRACT_MGR",
       "PROBITY",
       "FINANCE",
       "ADMIN",
       "EXEC",
       "SUPPLIER"
      ]
     },
     "orgUnit": {
      "type": "string"
     }
    },
    "additionalProperties": false,
    "required": [
     "name",
     "email",
     "role"
    ]
   },
...
  },
  "responses": {
   "E400": {
    "description": "Validation error",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E401": {
    "description": "Not authenticated",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E403": {
    "description": "Forbidden",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E404": {
    "description": "Not found",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E409": {
    "description": "Conflict / rule violated",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E422": {
    "description": "Business rule failed",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
...
  }
 }
}